requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1